+++ package.json
@@ -10,24 +10,23 @@
     "html-entities": "^1.2.0",
     "matchdep": "^1.0.0",
     "mustache": "^2.2.1",
-    "del": "^2.2.0"
-  },
-  "devDependencies": {
+    "del": "^2.2.0",
+    "browser-sync": "^2.11.1",
     "bs-html-injector": "^3.0.0",
     "grunt": "~0.4.5",
-    "grunt-contrib-concat": "^0.5.1",
     "grunt-browser-sync": "^2.2.0",
+    "grunt-contrib-concat": "^0.5.1",
     "grunt-contrib-copy": "^0.8.2",
-    "grunt-contrib-nodeunit": "^0.4.1",
     "grunt-contrib-watch": "^0.6.1",
-    "browser-sync": "^2.11.1",
     "gulp": "^3.9.0",
     "gulp-connect": "^2.3.1",
     "gulp-copy": "0.0.2",
     "gulp-header": "^1.7.1",
     "gulp-load": "^0.1.1",
-    "gulp-nodeunit": "0.0.5",
     "gulp-strip-banner": "0.0.2"
+  },
+  "devDependencies": {
+    "nodeunit": "^0.9.1"
   },
   "keywords": [
     "Pattern Lab",
@@ -45,7 +44,7 @@
   "author": "Brian Muenzenmeyer",
   "license": "MIT",
   "scripts": {
-    "test": "grunt travis --verbose"
+    "test": "node_modules/nodeunit/bin/nodeunit test"
   },
   "engines": {
     "node": ">=0.1"
